displayaudio: Update NxBTAudio
<libappbt.so>
This patch adds a GKI_delay 300ms (BSA recommendation) for smooth avrc command transmission.

<libappbt.so>
This patch fixes issue launcher is terminated when switching audio focus
bt audio (alsa closing) to local music player.

<libnxbt.so>
This patch adds exception handling for cases where the remote device name does not exist
when connecting to the AVK intermittently.

<Frame.cpp>
This patch changes the condition mapping of the playing UI from the play start action to a streaming start callback event.
This is because if the play start avrc command is transmitted but the BSA is not ready yet,
the pause event is called back and the streaming starts when the preparation is completed.
In order to eliminate the phenomenon of unnecessarily toggling the play status UI in this process.

<Frame.cpp>
This patch fixes the problem that restarting streaming when running NxBTAudio app
when A2DP streaming is being performed on remote device.

<NxBTService.cpp>
This patch changes the status token in the streaming start event callback stub function.
To be precise, Not open failure, it is not opened status.

Change-Id: I62da90755c8b8aa34a1ee0576012ca789884f43f
Signed-off-by: Lee GeunJeong <kjlee@nexell.co.kr>